Fix lock_manager, improve procedure of unzip, align option style

+++ unzip.py
@@ -1,33 +1,31 @@
 import os
 from pathlib import Path
-import shutil
 import tempfile
 import zipfile
 
 from flowweave import FlowWeaveResult
-
 from .file_system import FileSystem
 from .lock_manager import get_path_lock
 
 class Unzip(FileSystem):
     def operation_init(self):
-        self.zips = None
+        self.unzip = {}
 
     def operation(self):
         result = FlowWeaveResult.SUCCESS
         self.message(f"source : {self.source_dir}")
-        self.message(f"export : {self.export_dir}")
 
-        if self.zips:
-            zips = self.zips if isinstance(self.zips, list) else [self.zips]
+        zips = self.unzip.get("zips")
+        if zips:
+            zips = zips if isinstance(zips, list) else [zips]
         else:
             zips = self.prev_future.get("data", {}).get("zips", [])
 
         for source_dir in self.source_dir:
             for export_dir in self.export_dir:
                 for zip in zips:
-                    dst_root = self.unzip_file_from_source(source_dir, export_dir, zip)
-                    self.message(f"unzip : {dst_root}")
+                    self.message(f"UnZip: {zip} - {source_dir} -> {export_dir}")
+                    self.unzip_file_from_source(source_dir, export_dir, zip)
 
         return result
 
@@ -39,37 +37,31 @@ class Unzip(FileSystem):
                 raise RuntimeError(f"Unsafe path in zip: {member.filename}")
             zf.extract(member, dest)
 
-    def _remove_path(p: Path):
-        if p.is_dir():
-            shutil.rmtree(p, ignore_errors=False)
-        else:
-            p.unlink(missing_ok=True)
-
     def unzip_file_from_source(self, source_dir: str, export_dir: str, zip_name: str) -> Path:
         src_root = Path(source_dir).resolve()
         dst_root = Path(export_dir).resolve()
-
         zip_path = src_root / zip_name
+
         if not zip_path.exists():
-            raise FileNotFoundError(f"{zip_path} not found")
+            raise FileNotFoundError(zip_path)
 
         dst_root.mkdir(parents=True, exist_ok=True)
-
         lock = get_path_lock(dst_root)
-
         with lock:
             with tempfile.TemporaryDirectory(dir=dst_root) as tmp_dir:
-                tmp_path = Path(tmp_dir)
+                tmp_root = Path(tmp_dir)
 
                 with zipfile.ZipFile(zip_path, "r") as zf:
-                    self._safe_extract(zf, tmp_path)
+                    self._safe_extract(zf, tmp_root)
 
-                for item in tmp_path.iterdir():
-                    target = dst_root / item.name
+                # merge
+                for item in tmp_root.rglob("*"):
+                    rel = item.relative_to(tmp_root)
+                    target = dst_root / rel
 
-                    if target.exists():
-                        self._remove_path(target)
+                    if item.is_file():
+                        target.parent.mkdir(parents=True, exist_ok=True)
+                        os.replace(item, target)
 
-                    os.replace(item, target)
 
         return dst_root
